replace ``int ret`` by ``long long ret`` to avoid overflows

+++ c/reducercommutativity/avg20_true-unreach-call.c
@@ -18,7 +18,7 @@ extern void __VERIFIER_error() __attribute__ ((__noreturn__));
 int avg (int x[N])
 {
   int i;
-  int ret;
+  long long ret;
   ret = 0;
   for (i = 0; i < N; i++) {
     ret = ret + x[i];
